requirements.txt
anytree
matplotlib==3.4.1
minds-kit
namedlist
numpy==1.20.0
orange3==3.31.1
pandas==1.3.4
pyrulelearn
pysmt
python-sat==0.1.8.dev2
scikit-learn==0.24.2
tqdm
wittgenstein
xgboost==1.2.1
z3-solver==4.8.17.0
